backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-multipart==0.0.6
pydantic==2.5.3
aiofiles==23.2.1
Pillow==10.2.0
pyvips==2.2.3
celery==5.3.6
redis==5.0.1
PyExifTool==0.5.6
httpx==0.27.2
numpy==1.26.4
opencv-python-headless==4.10.0.84
onnxruntime==1.20.1
insightface==0.7.3
